+++ src/Microsoft.AspNetCore.Mvc.Core/Internal/IControllerPropertyActivator.cs
@@ -1,7 +1,7 @@
 // Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-namespace Microsoft.AspNetCore.Mvc.Controllers
+namespace Microsoft.AspNetCore.Mvc.Internal
 {
     public interface IControllerPropertyActivator
     {
